LayoutTests/http/tests/security/mixedContent/insecure-plugin-in-iframe.html

<html>
<body>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
<p>This test loads a secure iframe that loads an insecure plugin.  We should
get a mixed content callback because the insecure plug-in can script the secure
origin.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-plugin.html"></iframe>
</body>
</html>
